Check existance before making nonexistant.

+++ install_emacs_config.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
-rm -r ~/.emacs.d/config
+[[ -e ~/.emacs.d/config ]] && rm -r ~/.emacs.d/config
 cp -r _emacs.d/* ~/.emacs.d
